+++ helper/error.js
@@ -59,7 +59,7 @@ const errorHandler = (check) => {
       logger.error(error)
       next(
         new GeneralError(
-          Messages.SERVER_ERROR,
+          Messages.INTERNAL_SERVER_ERROR,
           StatusCodes.INTERNAL_SERVER_ERROR,
           undefined,
           RESPONSE_STATUS.ERROR,
